Move CertForwarderMiddleware to HttpOverrides
